requirements.txt
numpy==1.16.3
pandas==0.24.2
scikit-learn==0.21.1
scipy==1.3.0
tensorflow_gpu>=1.13.1
keras>=2.2.4